move the logo from header to start page hero
