+++ src/Components/StartPageHero.tsx
@@ -5,9 +5,11 @@ import {
   InputBase,
   IconButton,
   SxProps,
+  Avatar,
 } from "@mui/material";
 import SearchIcon from "@mui/icons-material/Search";
 import React from "react";
+import logoImg from "../Assets/Images/logo.png";
 
 import mobileMainHeroImage from "../Assets/Images/mobileMainHeroImage.png";
 import MobileHeader from "../Assets/Images/mobileHeader.png";
@@ -76,7 +78,7 @@ const StartPageHero = () => {
                     }}
                     aria-label="search"
                   >
-                    <SearchIcon />
+                    <SearchIcon sx={{fill: '#FFFFFF'}}/>
                   </IconButton>
                 </Paper>
               </Box>
@@ -85,12 +87,29 @@ const StartPageHero = () => {
         </Box>
       </Box>
       <Box sx={desktopBox}>
+      <Box sx={mainLogoDiv}>
+          <Box sx={deskMainLogo}>
+            <Avatar alt="Logo" src={logoImg} sx={mainLogo} />
+            <Typography sx={mainDeskText}>
+              Lorem impsum dolor sit amet bla{" "}
+            </Typography>
+            <Typography sx={deskText}>
+              Borrow almost anything from people nearby for jobs at home, fun
+              experiences or work.
+            </Typography>
+          </Box>
+        </Box>
         <Box sx={DesktopMainHeroImgDiv}>
-          <img
-            src={DesktopMainHeroImg}
-            alt="Hero image"
-            style={{ width: "100%", height: "100%", margin: "auto" }}
-          />
+        <Box
+          component="img"
+          sx={{
+            width: "100%", 
+            height: "100%", 
+            margin: "auto"
+          }}
+          alt="Hero image"
+          src={DesktopMainHeroImg}
+        />
         </Box>
       </Box>
     </>
@@ -121,10 +140,40 @@ const searchBox: SxProps = {
 const desktopBox: SxProps = {
   display: { xs: "none", md: "block", lg: "block", xl: "block" },
   userSelect: "none",
+ 
+};
+const mainLogoDiv: SxProps = {
+  // width: '1
+  display: { xs: 'none', md:'flex', lg: 'flex', xl: 'flex' },
+  textAlign: 'center',
+  width: '100%',
+  justifyContent: 'center',
+  alignItems: 'center',
+}
+const deskMainLogo: SxProps = {
+  display: 'flex',
+  alignItems: 'center',
+  justifyContent: 'center',
+  flexDirection: 'column',
+  mt: {xs: 'none', md: '5em', lg: '0em', xl: '0em'}
+}
+const mainLogo: SxProps = {
+  width: "200px",
+  height: "200px",
+};
+const mainDeskText: SxProps = {
+  color: "black",
+  marginTop: "1em",
+  fontSize: "1.5em",
+  fontweight: "bold",
+};
+const deskText: SxProps = {
+  color: "black",
 };
 const DesktopMainHeroImgDiv: SxProps = {
   // float: 'right',
-  width: "90%",
+  width: "50%",
   margin: "auto",
+  mt: '2em',
 };
 export default StartPageHero;
